# backend/src/handlers.rs
use axum::{
    extract::{Path, Query, State},
    response::IntoResponse,
    Json,
};
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use std::sync::atomic::Ordering;
use std::sync::Arc;
use std::time::{Instant, SystemTime, UNIX_EPOCH};
use uuid::Uuid;

use crate::api::corridors::ListCorridorsQuery;
use crate::broadcast::{broadcast_anchor_update, broadcast_corridor_update};
use crate::cache::CacheManager;
use crate::database::Database;
use crate::error::{ApiError, ApiResult};
use crate::models::corridor::Corridor;
use crate::models::{CreateAnchorRequest, CreateCorridorRequest};
use crate::rpc::StellarRpcClient;
use crate::state::AppState;

pub mod job_monitoring;
/// DTO for corridor transaction data
#[derive(Debug, Deserialize, Clone)]
pub struct CorridorTransactionDto {
    pub status: String,
    pub settlement_time_ms: i64,
}

#[derive(Serialize, Debug, Clone)]
#[serde(rename_all = "camelCase")]
pub struct HealthStatus {
    pub status: String,
    pub timestamp: DateTime<Utc>,
    pub version: String,
    pub uptime_seconds: u64,
    pub checks: HealthChecks,
    pub network: NetworkInfo,
    pub client_info: Option<ClientInfo>,
}

#[derive(Serialize, Debug, Clone)]
pub struct NetworkInfo {
    pub network: String,
    pub display_name: String,
    pub is_primary: bool,
}

#[derive(Serialize, Debug, Clone)]
pub struct ClientInfo {
    pub client_type: String,
    pub client_version: Option<String>,
}

#[derive(Serialize, Debug, Clone)]
pub struct HealthChecks {
    pub database: ComponentHealth,
    pub cache: ComponentHealth,
    pub rpc: ComponentHealth,
}

#[derive(Serialize, Debug, Clone)]
pub struct ComponentHealth {
    pub healthy: bool,
    pub response_time_ms: Option<u64>,
    pub message: Option<String>,
}

#[derive(Serialize)]
pub struct ListCorridorsResponse {
    pub corridors: Vec<Corridor>,
    pub total: usize,
}

/// Check database health
async fn check_database(db: &Arc<Database>) -> ComponentHealth {
    let start = Instant::now();
    match sqlx::query("SELECT 1").fetch_one(db.pool()).await {
        Ok(_) => ComponentHealth {
            healthy: true,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: None,
        },
        Err(e) => ComponentHealth {
            healthy: false,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: Some(format!("Database connection failed: {}", e)),
        },
    }
}

/// Check cache health
async fn check_cache(cache: &Arc<CacheManager>) -> ComponentHealth {
    let start = Instant::now();
    match cache.ping().await {
        Ok(_) => ComponentHealth {
            healthy: true,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: None,
        },
        Err(e) => ComponentHealth {
            healthy: false,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: Some(format!("Cache connection failed: {}", e)),
        },
    }
}

/// Check RPC health
async fn check_rpc(rpc: &Arc<StellarRpcClient>) -> ComponentHealth {
    let start = Instant::now();
    match rpc.check_health().await {
        Ok(_) => ComponentHealth {
            healthy: true,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: None,
        },
        Err(e) => ComponentHealth {
            healthy: false,
            response_time_ms: Some(start.elapsed().as_millis() as u64),
            message: Some(format!("RPC connection failed: {}", e)),
        },
    }
}

/// Detailed health check endpoint with network and client context
pub async fn health_check(State(app_state): State<AppState>) -> Json<HealthStatus> {
    let db_health = check_database(&app_state.db).await;
    let cache_health = check_cache(&app_state.cache).await;
    let rpc_health = check_rpc(&app_state.rpc_client).await;

    let overall_status = if db_health.healthy && cache_health.healthy && rpc_health.healthy {
        "healthy"
    } else if db_health.healthy && cache_health.healthy {
        "degraded"
    } else {
        "unhealthy"
    };

    let start_epoch = app_state.server_start_time.load(Ordering::Relaxed);
    let now_epoch = SystemTime::now()
        .duration_since(UNIX_EPOCH)
        .map_or(0, |d| d.as_secs());
    let uptime_seconds = now_epoch.saturating_sub(start_epoch);

    let network_context = &app_state.network_context;
    let is_primary = network_context.network == app_state.multi_network_config.primary_network;

    Json(HealthStatus {
        status: overall_status.to_string(),
        timestamp: Utc::now(),
        version: env!("CARGO_PKG_VERSION").to_string(),
        uptime_seconds,
        checks: HealthChecks {
            database: db_health,
            cache: cache_health,
            rpc: rpc_health,
        },
        network: NetworkInfo {
            network: network_context.network_id(),
            display_name: network_context.display_name().to_string(),
            is_primary,
        },
        client_info: None,
    })
}

/// PUT /api/anchors/:id/metrics - Update anchor metrics
#[derive(Debug, Deserialize)]
pub struct UpdateMetricsRequest {
    pub total_transactions: i64,
    pub successful_transactions: i64,
    pub failed_transactions: i64,
    pub avg_settlement_time_ms: Option<i32>,
    pub volume_usd: Option<f64>,
}

pub async fn update_anchor_metrics(
    State(app_state): State<AppState>,
    Path(id): Path<Uuid>,
    Json(req): Json<UpdateMetricsRequest>,
) -> ApiResult<Json<crate::models::Anchor>> {
    // Verify anchor exists
    if app_state.db.get_anchor_by_id(id).await?.is_none() {
        return Err(ApiError::not_found(
            "ANCHOR_NOT_FOUND",
            format!("Anchor with id {id} not found"),
        ));
    }

    let computed = crate::analytics::compute_anchor_metrics(
        req.total_transactions,
        req.successful_transactions,
        req.failed_transactions,
        req.avg_settlement_time_ms,
    );
    let anchor = app_state
        .db
        .update_anchor_metrics(crate::database::AnchorMetricsUpdate {
            anchor_id: id,
            total_transactions: req.total_transactions,
            successful_transactions: req.successful_transactions,
            failed_transactions: req.failed_transactions,
            avg_settlement_time_ms: req.avg_settlement_time_ms,
            volume_usd: req.volume_usd,
            reliability_score: computed.reliability_score,
            success_rate: computed.success_rate,
            failure_rate: computed.failure_rate,
            status: computed.status.as_str().to_string(),
        })
        .await?;

    // Broadcast the anchor update to WebSocket clients
    broadcast_anchor_update(&app_state.ws_state, &anchor);

    Ok(Json(anchor))
}

/// GET /api/anchors/:id/assets - Get assets for an anchor
pub async fn get_anchor_assets(
    State(app_state): State<AppState>,
    Path(id): Path<Uuid>,
) -> ApiResult<Json<Vec<crate::models::Asset>>> {
    // Verify anchor exists
    if app_state.db.get_anchor_by_id(id).await?.is_none() {
        return Err(ApiError::not_found(
            "ANCHOR_NOT_FOUND",
            format!("Anchor with id {id} not found"),
        ));
    }

    let assets = app_state.db.get_assets_by_anchor(id).await?;

    Ok(Json(assets))
}

/// POST /api/anchors/:id/assets - Add asset to anchor
#[derive(Debug, Deserialize)]
pub struct CreateAssetRequest {
    pub asset_code: String,
    pub asset_issuer: String,
}

pub async fn create_anchor_asset(
    State(app_state): State<AppState>,
    Path(id): Path<Uuid>,
    Json(req): Json<CreateAssetRequest>,
) -> ApiResult<Json<crate::models::Asset>> {
    // Verify anchor exists
    if app_state.db.get_anchor_by_id(id).await?.is_none() {
        return Err(ApiError::not_found(
            "ANCHOR_NOT_FOUND",
            format!("Anchor with id {id} not found"),
        ));
    }

    let asset = app_state
        .db
        .create_asset(id, req.asset_code, req.asset_issuer)
        .await?;

    Ok(Json(asset))
}

/// GET /api/anchors/:id - Get anchor by ID
pub async fn get_anchor(
    State(app_state): State<AppState>,
    Path(id): Path<Uuid>,
) -> ApiResult<Json<crate::models::Anchor>> {
    let anchor =
        app_state.db.get_anchor_by_id(id).await?.ok_or_else(|| {
            ApiError::not_found("ANCHOR_NOT_FOUND", format!("Anchor {id} not found"))
        })?;
    Ok(Json(anchor))
}

/// GET /api/anchors/account/:stellar_account - Get anchor by Stellar account
pub async fn get_anchor_by_account(
    State(app_state): State<AppState>,
    Path(stellar_account): Path<String>,
) -> ApiResult<Json<crate::models::Anchor>> {
    let anchor = app_state
        .db
        .get_anchor_by_stellar_account(&stellar_account)
        .await?
        .ok_or_else(|| {
            ApiError::not_found(
                "ANCHOR_NOT_FOUND",
                format!("Anchor for account {stellar_account} not found"),
            )
        })?;
    Ok(Json(anchor))
}

/// GET /metrics - Prometheus metrics endpoint (all registered metrics via global registry)
pub async fn get_prometheus_metrics() -> impl IntoResponse {
    crate::observability::metrics::metrics_handler()
}

#[cfg(test)]
fn render_pool_metrics_prometheus(metrics: &crate::database::PoolMetrics) -> String {
    format!(
        "# HELP veltro_db_pool_size Database pool size\n\
# TYPE veltro_db_pool_size gauge\n\
veltro_db_pool_size {}\n\
# HELP veltro_db_pool_idle Database pool idle connections\n\
# TYPE veltro_db_pool_idle gauge\n\
veltro_db_pool_idle {}\n\
# HELP veltro_db_pool_active Database pool active connections\n\
# TYPE veltro_db_pool_active gauge\n\
veltro_db_pool_active {}\n",
        metrics.size, metrics.idle, metrics.active
    )
}

/// GET /db/pool-metrics - Return current database pool metrics as JSON
pub async fn pool_metrics(State(state): State<AppState>) -> impl IntoResponse {
    let metrics = state.db.pool_metrics();
    crate::observability::metrics::set_pool_size(metrics.size as i64);
    crate::observability::metrics::set_pool_idle(metrics.idle as i64);
    crate::observability::metrics::set_pool_active(metrics.active as i64);
    Json(metrics)
}

/// GET /api/corridors - List all corridors
pub async fn list_corridors(
    State(app_state): State<AppState>,
    Query(params): Query<ListCorridorsQuery>,
) -> ApiResult<Json<ListCorridorsResponse>> {
    let corridors = app_state
        .db
        .list_corridors(params.limit, params.offset)
        .await?;
    let total = corridors.len();
    Ok(Json(ListCorridorsResponse { corridors, total }))
}

/// POST /api/anchors - Create a new anchor
pub async fn create_anchor(
    State(app_state): State<AppState>,
    Json(req): Json<CreateAnchorRequest>,
) -> ApiResult<Json<crate::models::Anchor>> {
    let anchor = app_state.db.create_anchor(req).await?;
    Ok(Json(anchor))
}

/// POST /api/corridors - Create a new corridor
pub async fn create_corridor(
    State(app_state): State<AppState>,
    Json(req): Json<CreateCorridorRequest>,
) -> ApiResult<Json<Corridor>> {
    let corridor = app_state.db.create_corridor(req).await?;
    // Broadcast the new corridor to WebSocket clients
    broadcast_corridor_update(&app_state.ws_state, &corridor);
    Ok(Json(corridor))
}

/// PUT /api/corridors/:id/metrics-from-transactions - Compute metrics from transactions and persist
#[derive(Debug, Deserialize)]
pub struct UpdateCorridorMetricsFromTxns {
    pub transactions: Vec<CorridorTransactionDto>,
}

/// PUT /api/corridors/:id/metrics-from-transactions - Placeholder for updating metrics from batch transactions
pub async fn update_corridor_metrics_from_transactions(
    State(_app_state): State<AppState>,
    Path(_id): Path<Uuid>,
) -> ApiResult<Json<serde_json::Value>> {
    // Implementation for processing transaction batch logic goes here
    Ok(Json(serde_json::json!({ "status": "not_implemented" })))
}

pub async fn ingestion_status(
    State(app_state): State<AppState>,
) -> ApiResult<Json<crate::ingestion::IngestionStatus>> {
    let status = app_state.ingestion.get_ingestion_status().await?;
    Ok(Json(status))
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_render_pool_metrics_prometheus() {
        let metrics = crate::database::PoolMetrics::new(12, 3, 9);
        let rendered = render_pool_metrics_prometheus(&metrics);

        assert!(rendered.contains("veltro_db_pool_size 12"));
        assert!(rendered.contains("veltro_db_pool_idle 3"));
        assert!(rendered.contains("veltro_db_pool_active 9"));
        assert!(rendered.contains("# TYPE veltro_db_pool_size gauge"));
    }
}