# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Veltro API",
    "description": "API for Stellar network analytics, anchor monitoring, and payment corridor insights",
    "contact": {
      "name": "Veltro Team",
      "email": "support@veltro.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.veltro.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/anchors/account/{stellar_account}": {
      "get": {
        "tags": [
          "Anchors"
        ],
        "summary": "GET /api/anchors/account/:stellar_account - Get anchor by Stellar account (G- or M-address)",
        "operationId": "get_anchor_by_account",
        "parameters": [
          {
            "name": "stellar_account",
            "in": "path",
            "description": "Anchor Stellar account address (G... or M...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anchor retrieved successfully"
          },
          "404": {
            "description": "Anchor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/anchors/{id}": {
      "get": {
        "tags": [
          "Anchors"
        ],
        "summary": "GET /api/anchors/:id - Get detailed anchor information",
        "operationId": "get_anchor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Anchor UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anchor details retrieved successfully"
          },
          "404": {
            "description": "Anchor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/corridors": {
      "get": {
        "tags": [
          "Corridors"
        ],
        "summary": "List all payment corridors",
        "description": "Returns a list of payment corridors with performance metrics.\nSupports filtering by success rate, volume, and asset code.\n\n**DATA SOURCE: RPC**\n- Payment data from Horizon API\n- Trade data from Horizon API\n- Order book data from Horizon API\n- Calculates corridor metrics from real-time RPC data",
        "operationId": "list_corridors",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (default: 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 0
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort by field (`success_rate` or volume)",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortBy"
            }
          },
          {
            "name": "success_rate_min",
            "in": "query",
            "description": "Minimum success rate filter",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 95.0
          },
          {
            "name": "success_rate_max",
            "in": "query",
            "description": "Maximum success rate filter",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 100.0
          },
          {
            "name": "volume_min",
            "in": "query",
            "description": "Minimum volume filter (USD)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 100000.0
          },
          {
            "name": "volume_max",
            "in": "query",
            "description": "Maximum volume filter (USD)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 10000000.0
          },
          {
            "name": "asset_code",
            "in": "query",
            "description": "Filter by asset code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "USDC"
          },
          {
            "name": "time_period",
            "in": "query",
            "description": "Time period for metrics (24h, 7d, 30d)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "24h"
          }
        ],
        "responses": {
          "200": {
            "description": "List of corridors retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CorridorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorridorResponse": {
        "type": "object",
        "required": [
          "id",
          "source_asset",
          "destination_asset",
          "success_rate",
          "total_attempts",
          "successful_payments",
          "failed_payments",
          "average_latency_ms",
          "median_latency_ms",
          "p95_latency_ms",
          "p99_latency_ms",
          "liquidity_depth_usd",
          "liquidity_volume_24h_usd",
          "liquidity_trend",
          "health_score",
          "last_updated"
        ],
        "properties": {
          "average_latency_ms": {
            "type": "number",
            "format": "double",
            "description": "Average latency in milliseconds",
            "example": 450.5
          },
          "destination_asset": {
            "type": "string",
            "description": "Destination asset code",
            "example": "XLM"
          },
          "failed_payments": {
            "type": "integer",
            "format": "int64",
            "description": "Number of failed payments",
            "example": 10
          },
          "health_score": {
            "type": "number",
            "format": "double",
            "description": "Overall health score (0-100)",
            "example": 95.5
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for the corridor",
            "example": "USDC:native->XLM:native"
          },
          "last_updated": {
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "liquidity_depth_usd": {
            "type": "number",
            "format": "double",
            "description": "Liquidity depth in USD",
            "example": 1500000.0
          },
          "liquidity_trend": {
            "type": "string",
            "description": "Liquidity trend (increasing, stable, decreasing)",
            "example": "stable"
          },
          "liquidity_volume_24h_usd": {
            "type": "number",
            "format": "double",
            "description": "24-hour trading volume in USD",
            "example": 150000.0
          },
          "median_latency_ms": {
            "type": "number",
            "format": "double",
            "description": "Median latency in milliseconds",
            "example": 380.0
          },
          "p95_latency_ms": {
            "type": "number",
            "format": "double",
            "description": "95th percentile latency in milliseconds",
            "example": 850.0
          },
          "p99_latency_ms": {
            "type": "number",
            "format": "double",
            "description": "99th percentile latency in milliseconds",
            "example": 1200.0
          },
          "source_asset": {
            "type": "string",
            "description": "Source asset code",
            "example": "USDC"
          },
          "success_rate": {
            "type": "number",
            "format": "double",
            "description": "Success rate percentage",
            "example": 99.8
          },
          "successful_payments": {
            "type": "integer",
            "format": "int64",
            "description": "Number of successful payments",
            "example": 4990
          },
          "total_attempts": {
            "type": "integer",
            "format": "int64",
            "description": "Total payment attempts",
            "example": 5000
          }
        }
      },
...
    }
  },
...
}